refactor!: revamp Miden SDK API and expose some modules;
