+++ sdk/base-sys/src/bindings/mod.rs
@@ -1,1 +1,6 @@
+pub mod account;
+pub mod note;
 pub mod tx;
+mod types;
+
+pub use types::*;
