+++ sdk/base-sys/src/lib.rs
@@ -3,5 +3,6 @@
 
 #[cfg(feature = "bindings")]
 pub mod bindings;
+
 #[cfg(feature = "masl-lib")]
 pub mod masl;
