+++ sdk/sdk/src/lib.rs
@@ -1,5 +1,5 @@
 #![no_std]
 
-pub use miden_base_sys::bindings::tx::*;
+pub use miden_base_sys::bindings::*;
 pub use miden_sdk_alloc::BumpAlloc;
 pub use miden_stdlib_sys::*;
