+++ sdk/stdlib-sys/src/lib.rs
@@ -3,5 +3,5 @@
 mod intrinsics;
 mod stdlib;
 
-pub use intrinsics::{felt::*, word::*, WordAligned};
+pub use intrinsics::*;
 pub use stdlib::*;
