+++ sdk/stdlib-sys/src/stdlib/crypto/hashes.rs
@@ -102,8 +102,7 @@ fn hash_1to1(
     input: [u8; 32],
     extern_hash_1to1: unsafe extern "C" fn(u32, u32, u32, u32, u32, u32, u32, u32, *mut u8),
 ) -> [u8; 32] {
-    use crate::WordAligned;
-
+    use crate::intrinsics::WordAligned;
     let input = unsafe { core::mem::transmute::<[u8; 32], [u32; 8]>(input) };
     unsafe {
         let mut ret_area = ::core::mem::MaybeUninit::<WordAligned<[u8; 32]>>::uninit();
