+++ sdk/stdlib-sys/src/stdlib/mod.rs
@@ -3,5 +3,3 @@ mod mem;
 
 pub use crypto::*;
 pub use mem::*;
-
-// TODO: simplify bindings like in tx_kernel
