+++ tests/integration/expected/abi_transform_tx_kernel_get_inputs_4.hir
@@ -10,7 +10,7 @@
         ;; Functions
         (func (export #entrypoint) (param i32)
             (block 0 (param v0 i32)
-                (call #miden_base_sys::bindings::tx::get_inputs v0)
+                (call #miden_base_sys::bindings::note::get_inputs v0)
                 (br (block 1)))
 
             (block 1
@@ -151,7 +151,7 @@
                 (br (block 7 v66)))
         )
 
-        (func (export #miden_base_sys::bindings::tx::get_inputs)
+        (func (export #miden_base_sys::bindings::note::get_inputs)
               (param i32)
             (block 0 (param v0 i32)
                 (let (v1 i32) (const.i32 0))
