+++ tests/integration/expected/rust_sdk/basic_wallet.hir
@@ -151,7 +151,7 @@
                 (store v25 v0)
                 (let (v26 i32) (const.i32 32))
                 (let (v27 i32) (add.wrapping v9 v26))
-                (call #miden_base_sys::bindings::tx::add_asset v27 v9)
+                (call #miden_base_sys::bindings::account::add_asset v27 v9)
                 (let (v28 (ptr i32)) (global.symbol #__stack_pointer))
                 (store v28 v5)
                 (br (block 1)))
@@ -231,7 +231,7 @@
                 (store v47 v6)
                 (let (v48 i32) (const.i32 64))
                 (let (v49 i32) (add.wrapping v15 v48))
-                (call #miden_base_sys::bindings::tx::remove_asset v49 v15)
+                (call #miden_base_sys::bindings::account::remove_asset v49 v15)
                 (let (v50 i32) (const.i32 64))
                 (let (v51 i32) (add.wrapping v15 v50))
                 (let (v52 i32) (const.i32 32))
@@ -676,7 +676,7 @@
                 (br (block 7 v66)))
         )
 
-        (func (export #miden_base_sys::bindings::tx::add_asset)
+        (func (export #miden_base_sys::bindings::account::add_asset)
               (param i32) (param i32)
             (block 0 (param v0 i32) (param v1 i32)
                 (let (v2 u32) (bitcast v1))
@@ -721,7 +721,7 @@
                 (ret))
         )
 
-        (func (export #miden_base_sys::bindings::tx::remove_asset)
+        (func (export #miden_base_sys::bindings::account::remove_asset)
               (param i32) (param i32)
             (block 0 (param v0 i32) (param v1 i32)
                 (let (v2 u32) (bitcast v1))
