+++ tests/integration/src/compiler_test.rs
@@ -789,6 +789,7 @@ impl CompilerTestBuilder {
                     r#"
                 #![no_std]
                 #![no_main]
+                #![allow(unused_imports)]
 
                 #[panic_handler]
                 fn my_panic(_info: &core::panic::PanicInfo) -> ! {{
@@ -803,6 +804,7 @@ impl CompilerTestBuilder {
                 use miden_stdlib_sys::*;
 
                 #[no_mangle]
+                #[allow(improper_ctypes_definitions)]
                 pub extern "C" fn entrypoint{}
             "#,
                     source
@@ -863,6 +865,7 @@ debug = true
                 format!(
                     r#"#![no_std]
 #![no_main]
+#![allow(unused_imports)]
 
 #[panic_handler]
 fn my_panic(_info: &core::panic::PanicInfo) -> ! {{
