--- tests/rust-apps-wasm/rust-sdk/basic-wallet/README.md
@@ -1,19 +1,0 @@
-# basic_wallet
-
-## Useful commands
-
-`basic_wallet` is built using the [Miden compiler](https://github.com/0xPolygonMiden/compiler).  
-
-`cargo miden` is a `cargo` cargo extension. Check out its [documentation](https://github.com/0xPolygonMiden/compiler/tree/main/tools/cargo-miden)
-for more details.
-
-
-### Build
-
-To build this project, run:
-
-```bash
-cargo miden build --release
-```
-
-The compiled Miden artifact will be located in `target/miden/release`.
