--- tests/rust-apps-wasm/rust-sdk/basic-wallet/rust-toolchain.toml
@@ -1,5 +1,0 @@
-[toolchain]
-channel = "nightly-2024-08-06"
-components = ["rustfmt", "rust-src"]
-targets = ["wasm32-wasi"]
-profile = "minimal"
